Move replace() to after we've established all imports
This fixes a bug where we'd be unable to find an `@import` after doing a
replace() with a blank file. Could not solve this without separating the
two functions.

Now we have a `while` loop while we find all of the matches and add them
to an object (`imports`) and then a for loop where we go through each
match and then `replace()` the code.